app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
clickhouse-connect>=0.8.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=3.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.50.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
websockets>=13.0
